+++ rusqlite/Changelog.md
@@ -1,3 +1,7 @@
+# Version UPCOMING (TBD)
+
+* Adds `backup` feature that exposes SQLite's online backup API.
+
 # Version 0.5.0 (2015-12-08)
 
 * Adds `trace` feature that allows the use of SQLite's logging, tracing, and profiling hooks.